WIP: add CSRF token to HTTP headers and submit with Ajax
