+++ reahl-web/reahl/web/fw.py
@@ -71,6 +71,7 @@ from reahl.component.exceptions import arg_checks
 from reahl.component.i18n import Catalogue
 from reahl.component.modelinterface import StandaloneFieldIndex, FieldIndex, Field, Event, ValidationConstraint,\
                                              Allowed, exposed, Event, Action
+from reahl.web.csrf import InvalidCSRFToken, CSRFToken
 
 _ = Catalogue('reahl-web')
 
@@ -2055,9 +2056,14 @@ class HeaderContent(Widget):
         self.page = page
 
     def render(self):
-        config = ExecutionContext.get_context().config
-        return ''.join([library.header_only_material(self.page)
-                        for library in config.web.frontend_libraries])
+        context = ExecutionContext.get_context()
+        token_string = context.session.get_csrf_token().as_signed_string()
+        csrf_meta = '<meta name="csrf-token" content="%s">' % token_string
+        config = context.config
+        library_header_material = ''.join([library.header_only_material(self.page)
+                                           for library in config.web.frontend_libraries])
+        return csrf_meta+library_header_material
+
     
 
 
@@ -2413,7 +2419,7 @@ class RemoteMethod(SubResource):
     sub_regex = 'method'
     sub_path_template = 'method'
 
-    def __init__(self, view, name, callable_object, default_result, idempotent=False, immutable=False, method=None):
+    def __init__(self, view, name, callable_object, default_result, idempotent=False, immutable=False, method=None, disable_csrf_check=False):
         super().__init__(view, name)
         self.idempotent = idempotent or immutable
         self.immutable = immutable
@@ -2422,6 +2428,7 @@ class RemoteMethod(SubResource):
         self.caught_exception = None
         self.input_values = None
         self.method = method
+        self.disable_csrf_check = disable_csrf_check
 
     @property
     def should_commit(self):
@@ -2451,7 +2458,9 @@ class RemoteMethod(SubResource):
            completed successfully."""
 
     def call_with_input(self, input_values, catch_exception):
-        context = ExecutionContext.get_context()
+        if not self.disable_csrf_check:
+            self.check_csrf_header()
+
         caught_exception = None
         return_value = None
         try:
@@ -2460,6 +2469,22 @@ class RemoteMethod(SubResource):
             self.caught_exception = caught_exception = ex
             self.input_values = input_values
         return (return_value, caught_exception)
+
+    def check_csrf_header(self):
+        context = ExecutionContext.get_context()
+
+        try:
+            received_token_string = context.request.headers['X-CSRF-TOKEN']
+        except KeyError:
+            raise HTTPForbidden()
+        try:
+            received_token = CSRFToken.from_coded_string(received_token_string)
+        except InvalidCSRFToken as ex:
+            raise HTTPForbidden()
+        if received_token.is_expired():
+            raise HTTPForbidden()
+        if not context.session.get_csrf_token().matches(received_token):
+            raise HTTPForbidden()
 
     def cleanup_after_transaction(self):
         super().cleanup_after_transaction()
@@ -2513,8 +2538,8 @@ class CheckedRemoteMethod(RemoteMethod):
        .. versionchanged:: 5.0
           Split immutable into immutable and idempotent kwargs.
     """
-    def __init__(self, view, name, callable_object, result, idempotent=False, immutable=False, **parameters):
-        super().__init__(view, name, callable_object, result, idempotent=idempotent, immutable=immutable)
+    def __init__(self, view, name, callable_object, result, idempotent=False, immutable=False, disable_csrf_check=False, **parameters):
+        super().__init__(view, name, callable_object, result, idempotent=idempotent, immutable=immutable, disable_csrf_check=disable_csrf_check)
         self.parameters = FieldIndex(self)
         for name, field in parameters.items():
             self.parameters.set(name, field)
@@ -2537,7 +2562,7 @@ class EventChannel(RemoteMethod):
        Programmers should not need to work with an EventChannel directly.
     """
     def __init__(self, form, controller, name):
-        super().__init__(form.view, name, self.delegate_event, None, idempotent=False, immutable=False)
+        super().__init__(form.view, name, self.delegate_event, None, idempotent=False, immutable=False, disable_csrf_check=True)
         self.controller = controller
         self.form = form
