+++ reahl-web/reahl/web/static/reahl.csrf.js
@@ -1,0 +1,33 @@
+/* Copyright 2021 Reahl Software Services (Pty) Ltd. All rights reserved. */
+/*
+    This file is part of Reahl.
+
+    Reahl is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as
+    published by the Free Software Foundation; version 3 of the License.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+
+(function($) {
+"use strict";
+    // From: https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#use-of-custom-request-headers
+    var csrf_token = $('meta[name="csrf-token"]').attr('content');
+
+    $.ajaxSetup({
+        beforeSend: function(xhr, settings) {
+            if (!this.crossDomain) {
+                xhr.setRequestHeader("X-CSRF-TOKEN", csrf_token);
+            }
+        }
+    });
+
+})(jQuery);
+
